tkorhon1: No ACTIVE_MESH anymore, evacuation_grid => evacuation_skip
No active_mesh array anymore in main.f90. The evacuation_skip is doing
now the job for the fire meshes before icyc<1.

Some evac.f90 related feature and bug fixes. Now the evacuation zones
are handled better in one evacuation mesh. Now the door selection algorithm
checks that the door is in the present evacuation zone. Also the read_evac,
the read_door is corrected. Now it has define_mesh=F for the "dummy doors"
that do not have to_node. This means that the input check for these doors
is now better, it does not check all things that are checked for real doors.


Former-commit-id: ef7d852f9c92464de56f1980ea17c6307ee5fc2e