tkorhon1: FDS Source, evacuation time change in EVAC_MAIN_LOOP.
Added line in EVAC_MAIN_LOOP so that at ICYC=1 T_EVAC is a little bit
less than T_FIRE. This way the EVAC_EFF file is written at ICYC=1.
This used to be the case before (when each fire mesh nad differend DT).
This was broken in the process, where same T and DT was made for all
fire meshes.


Former-commit-id: 7c5f8418304fecb7ba9a5450baceb476e8181910