tkorhon1: Evacuation changes in main.f90
Now main.f90 should work for evacuation with MPI. Results
of MPI and non-MPI are not yet exactly the same, but the
code runs nicely. Actually, fire results are not same,
i.e., MPI results are different from serial ones also.
This makes it harder to see, if MPI works now for evacuation
also. I need to do more tests.


Former-commit-id: c17687b49ddb2573ddc33fe191c78fc603f21ee6