Parse RAMP information from *.smv files and store.
SMV files contain the RAMP data that was specified in the FDS file,
however, this information is not parsed by Smokeview and is
currently left unused. This commit adds the case to the SMV parser
to parse this information and store it for future use.

This commit does not make use of the RAMP data in any way, just
parses and stores it.