separate image reading functions to separate file
Previously, the functions for reading images were in renderimage.c. This
commit moves the to readimage.c. While this is a cleaner logical
division, a more important difference is that the read image functions
can be used without depending on GLUT and GLUI (etc.).

As such they can now live in the shared directory and be used by all
programs.