split object functions and move to shared
There is a lot of useful code that is currently intertwined with GLUT
and GLUI. This splits that code out and put it into shared.

This introduces a new head `shared_structures.h` to hold some structs
which are used in shared files. This should eventually be factored out.

This is 3/7 patches that are required to parse and post-process *.smv files without GLUT/GLUI etc.