ci: vcpkg: disable default features of libgd
This commit does three things, firstly moves the vcpkg information (i.e.,
system libraries for Windows) from CI into the repo to ensure consistent
usage.

Secondly it disables features of libgd that are not used. One of these
features was withdrawn for security reasons by GitHub/Microsoft which
caused a build failure.

Thirdly it pins the vcpkg snapshot version so that any if any such
withdrawal happens again we'll get a warning from GitHub rather than
waiting for the build to fail.