add platform-specific function to get path of the running executable
This commit changes how the path of the running executable is
determined. Previously it would be taken from argv[0] which means either
passing argv around everywhere or storing it in a global variable (which
makes the code less flexible).

Technically, argv[0] could have issues as it's not guaranteed to be
correct. On Windows the path is relative which means doing more logic on
the path.

This commit replaces that logic with the dedicated API for each
platform:

  - Windows: GetModuleFileName
  - Linux: symlink at /proc/self/exe
  - MacOS: _NSGetExecutablePath

The API is changed to `char *GetBinDir()` and
`char *GetSmvRootPath()` and is never stored in a global.