add macro definition for pp_COMPVER
pp_COMPVER (i.e., what is printed as the compiler version) is defined in
the makefile by various scripts. ifdefs fall back to "unknown"
otherwise. This also means its not defined when using cmake.

This commit modifies the ifdefs to use compiler defined macros such as
__VERSION__. This means makes the scripts unnecessary.

This commit doesn't remove those scripts, but when they are not run a
complete compiler version will still be printed.