split colorbar functions into shared
There is a lot of useful code that is currently intertwined with GLUT
and GLUI. This splits that code out and put it into shared.

This introduces a new head shared_structures.h to hold some structs
which are used in shared files. This should eventually be factored out.

This does include some refactoring in that colorbar data is collected
into a struct of type `colorbar_collection` which can be used without
reference to global variables (including tests as shown in
`Tests/parse_colorbar.c`).