move path definitions to functions
Previously, all paths (including config paths and paths of data files)
were initialised into global variables or into a struct. This occured at
various points in the code and did not always use the same logic.

This commit places all of this logic into paths.c and uses functions to
create paths on demand. This should make the use of paths more
consistent. This will also help support things such as different working
directories. In many cases a lot of these paths are not used.

Whether fdsprefix or chidfilebase is used is preserved in this commit,
but it may be something worth revising as they can be different values.

This also abstracts away much of the buffer and directory separator
logic and defers to the underlying OS for that information.