use stderr for logging in shared
Generally, the shared code (e.g. dmalloc.c) uses stderr for logging,
but isn't consistent and sometimes stdout is used. This changes all of
uses of stdout for logging to stderr within shared/ (and one instance
in smokeview/).

This is particularly useful for programs (in particular smvq) that
outputs data to stdout, allowing us to separate logging from output
data.