switch from separate globals to global struct
This completes the earlier series of commits by switching simulation
data from disparate global variables into a single struct with type
smv_case.

Currently there is only a single global instance of this instance
(global_scase). This means there is no functional change.

Once this is changed functions can be used to take an smv_case as a
parameter, but this can be done on a function-by-function basis without
these large commits.