remove code for rendering objects with displaylists
Objects contained a "use_displaylist" option, however, the parser always
sets this to zero and the code is never triggered. This commit removes
"use_displaylist" and any code that branches on it.

The use of displaylists has been deprecated in OpenGL so it is unlikely
that this is a feature that will be added later.