win: add support for unicode files
On linux paths are bytestrings and UTF-8 is supported transparently. On
Windows, Unicode paths are a superset of UTF-16 and non-Unicode paths
can be any number of encodings. This means that Smokeview can have
issues opening files and directories with non-ASCII characters in their
path.

On Windows, a UNICODE_PATHS option has been added to take commandline
arguments as UTF-16 and convert them to UTF-8. Therefore, internally,
Smokeview treats everything as ASCII/UTF-8 just like it does on Linux.

Conversion functions from UTF-8 to UTF-16 have also been added and
included in the FOPEN, UNLINK, CHDIR, and MKDIR functions. Theses
functions are alsomoved to have more consistent definitions.

Tests are added to verify that this works correctly.