refactor ParseCommonOptions and remove globals
ParseCommonOptions parse some basic options which are common across all
executables (such as help and version). Previously, this stored the
results in global variables declared in shared/. This means that to use
any of the code in shared it was also necessary to set up this global
variables.

This refactors ParseCommonOptions to create a struct with these
arguments. These arguments are only used very locally to
ParseCommonOptions anyway so there was little benefit to having them
global.

As a side note, as of this commit there is not use of `INMAIN` or
`SVDECL` in shared/. which greatly helps with code reuse. Note the
simplification of the tests.