define default colors in shared
Smokeview defines a number of colors (nominally)as constants.
Previously, the INMAIN macro was used to copy the defintion across for
each executable.

This change places the definition of these colors within shared/ (and
can therefore be built in to a library).

It does not mark them as constant as that's not the way Smokeview uses
them, but they could be.

This (effectively) removes any use of the INMAIN macro in shared/ which
makes code reuse much easier.