remove circular dependency between shared and executables
There are a number of programs that are built in the smv repo and they
each depend on the code within shared/ (including smokeview itself).
This means shared/ is a library that each of these programs can link to.

Previously, there was a circular dependency. The code in smokeview/
dependend on the code in shared/ which in turn depended on the code in
smokeivew/. This also meant that the code in smokediff/ and smvq/
transitively depended on the code in smokeview/.

Earlier commits removed dependency on the built code so that shared/ and
each program could be built and linked separately. However, there
remained a circular dependency in the header files. For example: the
parsing code in readsmvfile.c depended on the definition of
TEMP_IGNITION_MAX which was defined in smokeview/smokeviewdefs.h.

This commit moves some definitions and includes such that the code in
shared/ now only depends on itself and external library code. Until now,
workarounds had been necessary.

A result of this is that pp_THREAD is set once for all programs, which
doesn't have a significant impact.