gif: don't set the render type when making an animated gif
render_filetype is used to control the format of rendered static images.
When making movies using the FFmpeg CLI it is necessary to set this to
produce the static images.

As the animated GIFs use GD, this is not necessary and might be
confusing as it would reset the default rendered images to GIF.